--- config.ini
@@ -1,4 +1,0 @@
-[Config]
-lang = en
-; Allow any float number
-breaktime = 5
